requirements.txt
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
python-dateutil==2.8.2
ta-lib==0.4.28
backtesting==0.5.0
python-binance==1.0.19
pytz 